ckd/requirements.txt
geopandas==1.0.1
joblib==1.4.2
numpy==2.0.2
pandas==2.2.2
pandas-datareader==0.10.0
pandas-gbq==0.28.0
pandas-stubs==2.2.2.240909
scikit-learn==1.6.1
sklearn-pandas==2.2.0